+++ core/aws/aws-sdk-php/src/Credentials/CredentialProvider.php
@@ -158,6 +158,11 @@ class CredentialProvider
                     }
                     // Refresh the result and forward the promise.
                     return $result = $provider();
+                })
+                ->otherwise(function($reason) use (&$result) {
+                    // Cleanup rejected promise.
+                    $result = null;
+                    return new Promise\RejectedPromise($reason);
                 });
         };
     }
@@ -289,7 +294,7 @@ class CredentialProvider
             if (!is_readable($filename)) {
                 return self::reject("Cannot read credentials from $filename");
             }
-            $data = parse_ini_file($filename, true, INI_SCANNER_RAW);
+            $data = \Aws\parse_ini_file($filename, true, INI_SCANNER_RAW);
             if ($data === false) {
                 return self::reject("Invalid credentials file: $filename");
             }