+++ core/cksource/ckfinder/src/CKSource/CKFinder/Acl/MaskBuilder.php
@@ -4,7 +4,7 @@
  * CKFinder
  * ========
  * https://ckeditor.com/ckeditor-4/ckfinder/
- * Copyright (c) 2007-2018, CKSource - Frederico Knabben. All rights reserved.
+ * Copyright (c) 2007-2019, CKSource - Frederico Knabben. All rights reserved.
  *
  * The software, this file and its contents are subject to the CKFinder
  * License. Please read the license.txt file before using, installing, copying,
@@ -19,8 +19,6 @@ namespace CKSource\CKFinder\Acl;
  *
  * A class used to build access control masks for folder access management.
  * Two masks are used to handle access rule inheritance from parent directories.
- *
- * @copyright 2016 CKSource - Frederico Knabben
  */
 class MaskBuilder
 {