+++ core/cksource/ckfinder/src/CKSource/CKFinder/Exception/AccessDeniedException.php
@@ -4,7 +4,7 @@
  * CKFinder
  * ========
  * https://ckeditor.com/ckeditor-4/ckfinder/
- * Copyright (c) 2007-2018, CKSource - Frederico Knabben. All rights reserved.
+ * Copyright (c) 2007-2019, CKSource - Frederico Knabben. All rights reserved.
  *
  * The software, this file and its contents are subject to the CKFinder
  * License. Please read the license.txt file before using, installing, copying,
@@ -22,8 +22,6 @@ use Symfony\Component\HttpFoundation\Response;
  *
  * Thrown when file system permissions do not allow to perform an operation
  * such as accessing a directory or writing a file.
- *
- * @copyright 2016 CKSource - Frederico Knabben
  */
 class AccessDeniedException extends CKFinderException
 {