+++ core/cksource/ckfinder/src/CKSource/CKFinder/Exception/InvalidUploadException.php
@@ -4,7 +4,7 @@
  * CKFinder
  * ========
  * https://ckeditor.com/ckeditor-4/ckfinder/
- * Copyright (c) 2007-2018, CKSource - Frederico Knabben. All rights reserved.
+ * Copyright (c) 2007-2019, CKSource - Frederico Knabben. All rights reserved.
  *
  * The software, this file and its contents are subject to the CKFinder
  * License. Please read the license.txt file before using, installing, copying,
@@ -20,8 +20,6 @@ use CKSource\CKFinder\Error;
  * The "invalid upload" exception class.
  *
  * Thrown when an invalid file upload request was received.
- *
- * @copyright 2016 CKSource - Frederico Knabben
  */
 class InvalidUploadException extends CKFinderException
 {