+++ core/cksource/ckfinder/src/CKSource/CKFinder/Exception/MethodNotAllowedException.php
@@ -4,7 +4,7 @@
  * CKFinder
  * ========
  * https://ckeditor.com/ckeditor-4/ckfinder/
- * Copyright (c) 2007-2018, CKSource - Frederico Knabben. All rights reserved.
+ * Copyright (c) 2007-2019, CKSource - Frederico Knabben. All rights reserved.
  *
  * The software, this file and its contents are subject to the CKFinder
  * License. Please read the license.txt file before using, installing, copying,
@@ -21,8 +21,6 @@ use Symfony\Component\HttpFoundation\Response;
  * The "invalid HTTP method" exception class.
  *
  * Thrown when a command is called using the invalid HTTP method.
- *
- * @copyright 2016 CKSource - Frederico Knabben
  */
 class MethodNotAllowedException extends CKFinderException
 {