+++ core/cksource/ckfinder/src/CKSource/CKFinder/ExceptionHandler.php
@@ -4,7 +4,7 @@
  * CKFinder
  * ========
  * https://ckeditor.com/ckeditor-4/ckfinder/
- * Copyright (c) 2007-2018, CKSource - Frederico Knabben. All rights reserved.
+ * Copyright (c) 2007-2019, CKSource - Frederico Knabben. All rights reserved.
  *
  * The software, this file and its contents are subject to the CKFinder
  * License. Please read the license.txt file before using, installing, copying,
@@ -27,8 +27,6 @@ use \Symfony\Component\HttpKernel\Exception\HttpException;
  * The exception handler class.
  *
  * All errors are resolved here and passed to the response.
- *
- * @copyright 2016 CKSource - Frederico Knabben
  */
 class ExceptionHandler implements EventSubscriberInterface
 {