+++ core/cksource/ckfinder/src/CKSource/CKFinder/Filesystem/File/File.php
@@ -4,7 +4,7 @@
  * CKFinder
  * ========
  * https://ckeditor.com/ckeditor-4/ckfinder/
- * Copyright (c) 2007-2018, CKSource - Frederico Knabben. All rights reserved.
+ * Copyright (c) 2007-2019, CKSource - Frederico Knabben. All rights reserved.
  *
  * The software, this file and its contents are subject to the CKFinder
  * License. Please read the license.txt file before using, installing, copying,
@@ -144,10 +144,8 @@ abstract class File
         $i = 0;
         while (true) {
             $i++;
-            // ---------------------------------------------------------------------------------------
-            // 自动重命名 AutoRename
-            $this->fileName = "{$basename}-{$i}.{$extension}";
-            // ---------------------------------------------------------------------------------------
+            $this->fileName = "{$basename}({$i}).{$extension}";
+
             $filePath = Path::combine($path, $this->fileName);
 
             if (!$backend->has($filePath)) {