+++ core/cksource/ckfinder/src/CKSource/CKFinder/ResizedImage/ResizedImageRepository.php
@@ -4,7 +4,7 @@
  * CKFinder
  * ========
  * https://ckeditor.com/ckeditor-4/ckfinder/
- * Copyright (c) 2007-2018, CKSource - Frederico Knabben. All rights reserved.
+ * Copyright (c) 2007-2019, CKSource - Frederico Knabben. All rights reserved.
  *
  * The software, this file and its contents are subject to the CKFinder
  * License. Please read the license.txt file before using, installing, copying,
@@ -31,8 +31,6 @@ use CKSource\CKFinder\ResourceType\ResourceType;
  * A class responsible for resized image management that simplifies
  * operations on resized versions of the image file, like batch renaming/moving
  * together with the original file.
- *
- * @copyright 2016 CKSource - Frederico Knabben
  */
 class ResizedImageRepository
 {