+++ core/cksource/ckfinder/src/CKSource/CKFinder/Response/JsonResponse.php
@@ -4,7 +4,7 @@
  * CKFinder
  * ========
  * https://ckeditor.com/ckeditor-4/ckfinder/
- * Copyright (c) 2007-2018, CKSource - Frederico Knabben. All rights reserved.
+ * Copyright (c) 2007-2019, CKSource - Frederico Knabben. All rights reserved.
  *
  * The software, this file and its contents are subject to the CKFinder
  * License. Please read the license.txt file before using, installing, copying,
@@ -18,8 +18,6 @@ use Symfony\Component\HttpFoundation;
 
 /**
  * The CKFinder JSON response class.
- *
- * @copyright 2016 CKSource - Frederico Knabben
  */
 class JsonResponse extends HttpFoundation\JsonResponse
 {